# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: Country City API
  description: "An Spring Boot Rest API for managing countries and cities.<br />You\
    \ can find out more about filtering, paging and sorting at [github.com/furkanisitan/country-city-api](https://github.com/furkanisitan/country-city-api#filtering-paging-and-sorting)."
  contact:
    name: Furkan Işıtan
    url: https://furkanisitan.com/
    email: furkanisitan96@gmail.com
  license:
    name: MIT License
    url: https://raw.githubusercontent.com/furkanisitan/country-city-api/main/LICENSE
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
...
  /api/countries/{id}/cities:
    get:
      tags:
      - Countries
      summary: Returns a list of cities of the country.
      operationId: allCities
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              examples:
                no record:
                  description: "recordName: Country"
                  value:
                    success: false
                    message: The resource not found.
                    errors:
                    - "{recordName} not found for parameters (id='55')."
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultData'
              example:
                success: true
                message: The request has been processed successfully.
                payload:
                - id: 1
                  name: Ankara
                  population: 5747325
                  countryId: 1
components:
  schemas:
...
    Result:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        errors:
          type: array
          items:
            type: string
...
    ResultData:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        errors:
          type: array
          items:
            type: string
        payload:
          type: object
...